Feature ETP-621: Fix component library typescript errors.
