+++ packages/ComponentLibrary/src/components/FormView/selectors/QuantitySelector.tsx
@@ -1,11 +1,13 @@
 import React, { memo, useState, useCallback, useEffect } from 'react';
 import { TextField } from '@mui/material';
 import { QuantityProps } from '../types';
-import { validateNumber } from '../../../utils/quantitySelectorUtil';
+import { validateNumber } from '@workspaceui/componentlibrary/utils/quantitySelectorUtil';
 
 const INPUT_PROPS = {
-  inputMode: 'numeric' as const,
-  pattern: '[0-9]*',
+  inputProps: {
+    inputMode: 'numeric' as const,
+    pattern: '[0-9]*',
+  },
 };
 
 const QuantitySelector: React.FC<QuantityProps> = memo(
@@ -67,9 +69,7 @@ const QuantitySelector: React.FC<QuantityProps> = memo(
         error={error}
         helperText={error ? errorMessage : ' '}
         disabled={readOnly}
-        InputProps={{
-          inputProps: INPUT_PROPS,
-        }}
+        InputProps={INPUT_PROPS}
       />
     );
   },
