+++ packages/ComponentLibrary/src/components/FormView/selectors/SelectSelector.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from 'react';
 import Select from '@workspaceui/componentlibrary/components/Input/Select';
 import SearchOutlined from '../../../../ComponentLibrary/src/assets/icons/search.svg';
-import { topFilms } from '../../../../storybook/src/stories/Components/Input/Select/mock';
+import { topFilms } from '../../../../../storybook/src/stories/Components/Input/Select/mock';
 import { SelectSelectorProps } from '../types';
 import { useTheme } from '@mui/material';
