+++ packages/storybook/src/stories/Components/Table/types.ts
@@ -1,6 +1,7 @@
 import { Theme } from '@emotion/react';
 import { SxProps } from '@mui/material';
-import { FieldDefinition, BaseFieldDefinition, Section } from '../../../../../MainUI/screens/Form/types';
+import { FieldDefinition, BaseFieldDefinition } from '../../../../../MainUI/screens/Form/types';
+import { Section } from '@workspaceui/componentlibrary/components/FormView/types';
 
 export type OrganizationField = FieldDefinition | Section;
